Packaging the project for setuptools
