+++ .gitignore
@@ -1,7 +1,24 @@
-*.pyc
-*.pyo
-*.pyd
+# Misc
 .DS_Store
-.idea
+
+# Project related folders
 weights
 gradio_cached_examples
+
+# Compilation and Distribution
+__pycache__
+_ext
+*.pyc
+*.pyd
+*.pyo
+*.so
+*.dll
+*.egg-info/
+build/
+dist/
+wheels/
+
+# Editor Settings
+.idea
+.vscode
+_darcs
