+++ fastsam/__init__.py
@@ -1,9 +1,1 @@
-# Ultralytics YOLO 🚀, AGPL-3.0 license
-
-from .model import FastSAM
-from .predict import FastSAMPredictor
-from .prompt import FastSAMPrompt
-# from .val import FastSAMValidator
-from .decoder import FastSAMDecoder
-
-__all__ = 'FastSAMPredictor', 'FastSAM', 'FastSAMPrompt', 'FastSAMDecoder'
+from .modeling import *
